Add environment file parsing for default tenant manager
