+++ sqs/tenants/defaultmanager/env_parser.go
@@ -1,0 +1,41 @@
+package defaultmanager
+
+import (
+	"bufio"
+	"fmt"
+	"os"
+	"strings"
+)
+
+func parseEnvFile() (map[string]string, error) {
+	envMap := make(map[string]string)
+
+	file, err := os.Open("/etc/.env")
+	if err != nil {
+		return nil, fmt.Errorf("error opening .env file: %v", err)
+	}
+	defer file.Close()
+
+	scanner := bufio.NewScanner(file)
+	for scanner.Scan() {
+		line := strings.TrimSpace(scanner.Text())
+		if line == "" || strings.HasPrefix(line, "#") {
+			continue
+		}
+
+		parts := strings.SplitN(line, "=", 2)
+		if len(parts) != 2 {
+			continue
+		}
+
+		key := strings.TrimSpace(parts[0])
+		value := strings.TrimSpace(parts[1])
+		envMap[key] = value
+	}
+
+	if err := scanner.Err(); err != nil {
+		return nil, fmt.Errorf("error reading .env file: %v", err)
+	}
+
+	return envMap, nil
+}
