+++ sqs/tenants/defaultmanager/manager.go
@@ -1,11 +1,8 @@
 package defaultmanager
 
 import (
-	"bufio"
 	"fmt"
-	"os"
 	"q/models"
-	"strings"
 )
 
 type DefaultTenantManager struct{}
@@ -35,37 +32,4 @@ func (tm *DefaultTenantManager) GetAWSSecretKey(accessKey string, region string)
 
 func NewDefaultTenantManager() models.TenantManager {
 	return &DefaultTenantManager{}
-}
-
-func parseEnvFile() (map[string]string, error) {
-	envMap := make(map[string]string)
-
-	file, err := os.Open("/etc/.env")
-	if err != nil {
-		return nil, fmt.Errorf("error opening .env file: %v", err)
-	}
-	defer file.Close()
-
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-		if line == "" || strings.HasPrefix(line, "#") {
-			continue
-		}
-
-		parts := strings.SplitN(line, "=", 2)
-		if len(parts) != 2 {
-			continue
-		}
-
-		key := strings.TrimSpace(parts[0])
-		value := strings.TrimSpace(parts[1])
-		envMap[key] = value
-	}
-
-	if err := scanner.Err(); err != nil {
-		return nil, fmt.Errorf("error reading .env file: %v", err)
-	}
-
-	return envMap, nil
 }
